Extracted date formatting logic from DateTimeString to a separate file.
